requirements.txt
BeautifulSoup==3.2.1
Django==1.4.2
MySQL-python==1.2.4c1
beautifulsoup4==4.1.3
django-tastypie==0.9.11
html5lib==0.95
mimeparse==0.1.3
python-dateutil==1.5
wsgiref==0.1.2
